Fix shell newline insertion
This method may be a bit slower, but it now takes into account the text which
already exists in the buffer.

Previously, we might insert a short line, then concatenate another short line on
to it, and so on until we end up with a long line. Now that won't happen.