Only start the server if there isn't one running
We were getting annoying popups when reloading our config, asking if we want to
kill the running server. We don't, since that would close all of our open
frames.

Taking a look at the server-start function, when we answer "No" (which we always
do), the function just quits and doesn't bother restarting anything. Hence it's
equivalent to not calling the function at all.

More importantly, the yes/no dialogue can pop up on remote displays which we
don't have access to, e.g. on a laptop using 'ssh -Y'. If that machine
disappears without giving an answer, Emacs can become unresponsive (its thread
seems to be stuck waiting for the non-existent answer).

This is actually quite common. For example, I have Emacs configured in daemon
mode as a system service, running on NixOS. When I update or change the system,
it tells Emacs to reload its config if any dependencies changed. Updates can
take a while, so they may be left overnight, with the Yes/No dialogue box
popping up in the middle of the night on one of various machines I've been using
that day. If that machine disconnects, or the battery dies, etc. then Emacs will
be hung.

This hopefully prevents that happening yet again.