Improve fill-column-indicator activation
We only want fci on graphical displays, but that's a problem if we're launching
Emacs daemon at boot. We start polling for a GUI in order to enable our desired
fonts, and we were hijacking that to enable fci, but that's a hack.

Instead, we now enable it globally, and we do the GUI check on a
buffer-by-buffer basis. As long as we're using a GUI when a file is opened, it
will get fci enabled. AFAIK we're not opening any files at startup (we open a
lot of shells, but those don't get fci anyway).