Move some programming-related packages to programming.el
If we want as much as possible to use use-package, it doesn't really make sense
to put all use-package declarations in one file.

In particular, I wanted to move some of the prelude keybindings into to magit
use-package declaration, so we can move away from prelude. However, they turn
out to conflict with a global keybinding, which the declarative style can't
override.

Hence I needed to add an imperative step to unbind this global key before using
use-package to declare the magit config. It doesn't make sense to put such
imperative keybinding stuff in a packages.el file, so I thought it better to
pull the magit use-package out into a programming.el; which I would've wanted to
do eventually anyway.

Whilst we're at it, we pull out a few other things, like scala-mode.