Add whitespace-cleanup-mode
This runs whitespace-cleanup when a buffer is saved (we enable it globally for
now). We used to just bind whitespace-cleanup to a before-save hook, but that
can be annoying if we're editing someone else's code, and they've left it full
of dodgy whitespace: when we used to save our edits to such files, the existing
dodgy whitespace would get cleaned up, and we'd get a massive git diff.

In contrast, whitespace-cleanup-mode performs a check when a buffer is first
opened. It runs whitespace-cleanup on a temporary copy of the buffer contents,
to see what would happen: if there's no change then the file is clean, so we're
safe to run whitespace-cleanup on save (it will only be cleaning up our mess);
if there is a change then the buffer must already be unclean, so
whitespace-cleanup will not be run, which keeps the existing cruft intact.