warbo-shells: adjust shx advice, to find existing bash binaries
The old advice was inflexible, and also pretty much obsolete now that we don't
need to use wrappedShell. It was hard-coding /bin/bash for remote hosts, to
avoid using /run/current-system/sw/bin/wrappedShell on them (which is unlikely
to exist).

That advice had a TODO saying it would be better to check what actually
exists. Well, that's what the current commit does (as well as avoiding
wrappedShell completely, since we can use normal bash binaries now). With this
change, using '#ssh foo' in shell-mode (with shx-mode enabled) will check for
the existence of files like "/ssh:foo:/run/current-system/sw/bin/bash",
"/ssh:foo:/usr/bin/bash", etc. and open the remote shell using the first one
that actually exists. In principle this could be used to look for wrappedShell,
and fall back to bash; but I'm moving away from wrappedShell, so didn't bother.

This new behaviour should make TRAMP shells work more seamlessly when the remote
end might be NixOS (which doesn't have /bin/bash or /usr/bin/bash, etc.); when
it's not NixOS, but the user has a ~/.nix-profile; for FHS distros with
/usr/bin/bash; for distros which eschew /usr in favour of /bin/bash; etc. As a
last resort it will fall back to /bin/sh, which almost always exists but is less
than ideal (e.g. it may be a different shell like Dash, or it may be in POSIX
compatibility mode, and it may not source files like .bashrc, etc.). Note that
if we do get dropped into a /bin/sh shell, we should probably 'exec bash' as the
first thing we do ;)

Note that this is a pretty convoluted alternative to, say, setting the default
shell via normal TRAMP configuration. However, this may be more robust when we
have to use IPs instead of host names (e.g. if we can't get mDNS working).