warbo-shells: hack for directory tracking with uppercase hostname
The directory tracking implemented by ansi-osc-directory-tracker is flawed,
since it relies on url-generic-parse-url, which "normalises" the hostname to be
lowercase. However, it then does a string= against (system-name), which won't
match if the latter contains some uppercase characters.

To work around this, we advise the (system-name) function to optionally
lowercase its output, depending on the variable
warbo-shell-want-lowercase-system-name, which is always nil except for when
we're calling ansi-osc-directory-tracker (again, using advice).