prelude-editor.el Fix advice when used non-interactively
The code inherited from Prelude seems to be advising a couple of functions with
a lambda that relies on (interactive) to get region bounds. However, that
doesn't work when called programmatically, causing "wrong number of arguments"
errors (specifically for 'untabify', called by 'whitespace-cleanup-mode').

This changes the advice to 'around', so that we can access the arguments
directly, and see whether they've been supplied or not. This avoids having to
rely on (interactive). We also pull the logic out into a named function, to make
debugging easier (it took a while to narrow-down the source of this problem,
based on an unhelpful backtrace).