Fix bytecode compilation
Trying to do this before (package-initialize) causes all sorts of problems, as
the compiler can't find the package files we're requiring. Since the whole point
of early-init/preload is to do stuff before (package-initialize), I've moved it
out.

Now we (a) load it just before our warbo package (this required extending the
in-personal macro to pass along extra arguments to use-package), and (b) we wrap
the logic in a function, so it can be given the paths rather than trying to
infer them.