editor-tests: Fix a bunch of tests for the yank advice
Turns out that we(/Prelude) were relying on transient-mark-mode to be active as
an Emacs default; however, that's not active when using --batch or --script,
which is how we're running the tests.

I also changed some of the tests to send keyboard events, rather than calling
Elisp functions directly (since I care about the keys; not what they're bound
to). That requires having the temp buffers selected in a window; so I added a
macro to do that.